refactor: extract question sets hook
